Crash Error on empty string fix
Fixed the error with the page crashing if you search with empty string.
The consol still logs an error:
"api.spotify.com/v1/search?q=&type=track:1 Failed to load resource: the server responded with a status of 400 ()"
Probably because of how error handling is set up in dataSource, maybe write errors to database instead?